LayoutTests/platform/ios/imported/w3c/web-platform-tests/mathml/relations/css-styling/floats/not-floating-001-expected.txt


FAIL float child ignored in maction assert_approx_equals: inline size expected 20 +/- 1 but got 0
FAIL float child ignored in menclose assert_approx_equals: inline size expected 94.21875 +/- 1 but got 38.84375
FAIL float child ignored in merror assert_approx_equals: inline size expected 85.0625 +/- 1 but got 29.6875
FAIL float child ignored in mfrac assert_approx_equals: inline size expected 22 +/- 1 but got 2
FAIL float child ignored in mi assert_approx_equals: block size expected 37 +/- 1 but got 29
FAIL float child ignored in mmultiscripts assert_approx_equals: inline size expected 40.625 +/- 1 but got 20
FAIL float child ignored in mn assert_approx_equals: block size expected 37 +/- 1 but got 29
FAIL float child ignored in mo assert_approx_equals: block size expected 37 +/- 1 but got 29
FAIL float child ignored in mover assert_approx_equals: inline size expected 20 +/- 1 but got 0
FAIL float child ignored in mpadded assert_approx_equals: inline size expected 83.0625 +/- 1 but got 27.6875
FAIL float child ignored in mphantom assert_approx_equals: inline size expected 83.0625 +/- 1 but got 27.6875
FAIL float child ignored in mroot assert_approx_equals: inline size expected 52.3125 +/- 1 but got 0
FAIL float child ignored in mrow assert_approx_equals: inline size expected 83.0625 +/- 1 but got 27.6875
FAIL float child ignored in ms assert_approx_equals: block size expected 37 +/- 1 but got 29
FAIL float child ignored in msqrt assert_approx_equals: inline size expected 101.890625 +/- 1 but got 46.515625
FAIL float child ignored in mstyle assert_approx_equals: inline size expected 83.0625 +/- 1 but got 27.6875
FAIL float child ignored in msub assert_approx_equals: inline size expected 40.625 +/- 1 but got 0
FAIL float child ignored in msubsup assert_approx_equals: inline size expected 40.625 +/- 1 but got 20
FAIL float child ignored in msup assert_approx_equals: inline size expected 40.625 +/- 1 but got 0
FAIL float child ignored in mtext assert_approx_equals: block size expected 37 +/- 1 but got 29
FAIL float child ignored in munder assert_approx_equals: inline size expected 20 +/- 1 but got 0
FAIL float child ignored in munderover assert_approx_equals: block size expected 67.75 +/- 1 but got 20
FAIL float child ignored in semantics assert_approx_equals: inline size expected 20 +/- 1 but got 0
